Remove analysis tests now that SonOfSirCachealot is working. Make db creation script more robust. Start adding detector information to block class.

+++ SonOfSirCachealot/create_db.sql
@@ -1,4 +1,4 @@
-DROP DATABASE [EDMDatabase]
+IF EXISTS(SELECT name FROM sys.databases WHERE name = 'EDMDatabase') DROP DATABASE [EDMDatabase]
 GO
 
 CREATE DATABASE EDMDatabase
